Parse reqs from html successfully
It was easier than expected when you have pen and paper
to jot down every single format the UBC Calendar can have.
The Calendar itself was probably auto-generated, after all.

This new algorithm actually clarified what CPSC 304's prereqs mean.
I thought it was either CPSC 221, or (CPSC 260 and EECE 320 and (etc.))
but it turned out to be (CPSC 221 or (CPSC 260 and EECE 320)) and etc.
And I spent so long manually parsing the prereqs too... );